+++ web-common/src/features/dashboards/leaderboard/LeaderboardControls.svelte
@@ -7,42 +7,19 @@
   import { crossfade, fly } from "svelte/transition";
   import { runtime } from "../../../runtime-client/runtime-store";
   import Spinner from "../../entity-management/Spinner.svelte";
-  import Delta from "@rilldata/web-common/components/icons/Delta.svelte";
-  import PieChart from "@rilldata/web-common/components/icons/PieChart.svelte";
-  import {
-    ButtonGroup,
-    SubButton,
-  } from "@rilldata/web-common/components/button-group";
 
-  import { useModelHasTimeSeries } from "@rilldata/web-common/features/dashboards/selectors";
   import {
     MetricsExplorerEntity,
     metricsExplorerStore,
   } from "../dashboard-stores";
   import { useMetaQuery } from "../selectors";
+  import LeaderboardContextColumnToggle from "./LeaderboardContextColumnToggle.svelte";
 
   export let metricViewName;
 
   $: metaQuery = useMetaQuery($runtime.instanceId, metricViewName);
 
   $: measures = $metaQuery.data?.measures;
-
-  $: hasTimeSeriesQuery = useModelHasTimeSeries(
-    $runtime.instanceId,
-    metricViewName
-  );
-  $: hasTimeSeries = $hasTimeSeriesQuery?.data;
-
-  $: {
-    console.log(
-      "hasTimeSeries",
-      hasTimeSeries,
-      "selectedTimeRange",
-      metricsExplorer.selectedTimeRange,
-      "selectedComparisonTimeRange",
-      metricsExplorer.selectedComparisonTimeRange
-    );
-  }
 
   let metricsExplorer: MetricsExplorerEntity;
   $: metricsExplorer = $metricsExplorerStore.entities[metricViewName];
@@ -92,6 +69,9 @@
   /** set the selection only if measures is not undefined */
   $: selection = measures ? activeLeaderboardMeasure : [];
 
+  $: validPercentOfTotal =
+    activeLeaderboardMeasure?.validPercentOfTotal || false;
+
   $: showHideDimensions = createShowHideDimensionsStore(
     metricViewName,
     metaQuery
@@ -105,53 +85,6 @@
   };
   const setAllDimensionsVisible = () => {
     showHideDimensions.setAllToVisible();
-  };
-
-  let disabledButtons: ("delta" | "pie")[] = [];
-  $: {
-    disabledButtons = [];
-    if (
-      !hasTimeSeries ||
-      metricsExplorer.selectedComparisonTimeRange === undefined
-    )
-      disabledButtons.push("delta");
-    if (activeLeaderboardMeasure?.validPercentOfTotal !== true)
-      disabledButtons.push("pie");
-  }
-
-  let selectedButton: "delta" | "pie" | null = null;
-  // NOTE: time comparison takes precedence over percent of total
-  $: selectedButton = metricsExplorer?.showComparison
-    ? "delta"
-    : metricsExplorer?.showPercentOfTotal
-    ? "pie"
-    : null;
-
-  const handleContextValueButtonGroupClick = (evt) => {
-    const value = evt.detail;
-    if (value === "delta" && selectedButton == "delta") {
-      metricsExplorerStore.displayComparison(metricViewName, false);
-    } else if (value === "delta" && selectedButton != "delta") {
-      metricsExplorerStore.displayComparison(metricViewName, true);
-    } else if (value === "pie" && selectedButton == "pie") {
-      metricsExplorerStore.displayPercentOfTotal(metricViewName, false);
-    } else if (value === "pie" && selectedButton != "pie") {
-      metricsExplorerStore.displayPercentOfTotal(metricViewName, true);
-    }
-  };
-
-  $: selectedButtons = selectedButton === null ? [] : [selectedButton];
-
-  const pieTooltips = {
-    selected: "Hide percent of total",
-    unselected: "Show percent of total",
-    disabled: "To show percent of total, show top values by a summable metric",
-  };
-
-  const deltaTooltips = {
-    selected: "Hide percent change",
-    unselected: "Show percent change",
-    disabled: "To show percent change, select a comparison period above",
   };
 </script>
 
@@ -187,19 +120,7 @@
       >
         <span class="font-bold truncate">{selection?.main}</span>
       </SelectMenu>
-
-      <ButtonGroup
-        selected={selectedButtons}
-        disabled={disabledButtons}
-        on:subbutton-click={handleContextValueButtonGroupClick}
-      >
-        <SubButton value={"delta"} tooltips={deltaTooltips}>
-          <Delta />%
-        </SubButton>
-        <SubButton value={"pie"} tooltips={pieTooltips}>
-          <PieChart />%
-        </SubButton>
-      </ButtonGroup>
+      <LeaderboardContextColumnToggle {metricViewName} {validPercentOfTotal} />
     </div>
   {:else}
     <div
